change version number to short

+++ CondFormats/RunInfo/interface/LuminosityInfo.h
@@ -8,7 +8,7 @@
  * each event will occur at one of these BX. BX is defined to be the number of the
  * bunch crossing where this event occurred.
  *
- * $Id: LuminosityInfo.h,v 1.3 2009/02/20 13:38:41 xiezhen Exp $
+ * $Id: LuminosityInfo.h,v 1.4 2009/02/23 10:40:04 xiezhen Exp $
  *
  ************************************************************/
  
@@ -53,7 +53,7 @@ namespace lumi{
     ///
     ///getter methods
     ///
-    int lumiVersionNumber()const;
+    short lumiVersionNumber()const;
     int lumisectionID()const;
     float deadTimeNormalization()const;
     size_t nBunchCrossing()const;
@@ -72,7 +72,7 @@ namespace lumi{
     ///
     ///setter methods. 
     ///
-    void setLumiVersionNumber(int versiondid);
+    void setLumiVersionNumber(short versiondid);
     void setLumiSectionId(int sectionid);
     void setDeadtimeNormalization(float dtimenorm);
     void setLumiAverage(const LumiAverage& avg,
@@ -84,7 +84,7 @@ namespace lumi{
     int m_sectionid; 
     float m_deadtime_normalization;
     std::vector<LumiAverage> m_summaryinfo;
-    int m_versionid;
+    short m_versionid;
   }; 
 }//ns lumi
 #endif
